Refactor and enhance article stock and money handling
Refactored article stock handling by introducing new classes and methods for querying and displaying article stock totals. Replaced `MoneyPad` and `PosMoneyPadDialog` with `InsertMoneyModal` for improved money handling. Added new UI components and modals for better user interaction. Updated various files to improve code readability and consistency.

+++ LogicPOS.Api/Features/Articles/Articles/Article.cs
@@ -54,8 +54,8 @@ namespace LogicPOS.Api.Entities
         public ArticlePrice Price5 { get; set; }
         public bool PriceWithVat { get; set; }
         public decimal Discount { get; set; }
-        public uint DefaultQuantity { get; set; }
-        public uint MinimumStock { get; set; }
+        public decimal DefaultQuantity { get; set; }
+        public decimal MinimumStock { get; set; }
         public decimal Tare { get; set; }
         public float Weight { get; set; }
         public string Barcode { get; set; }
